Internal: move FetchReader to util
Does not belong in randomAccessDataProviders which is a folder of data providers.
