+++ packages/studio-base/src/util/FetchReader.ts
